+++ packages/uiweb/src/lib/dataProviders/index.ts
@@ -1,1 +1,2 @@
 export * from "./SpaceDataProvider";
+export * from "./ChatDataProvider";