+++ packages/uiweb/src/lib/hooks/chat/useFetchHistoryMessages.ts
@@ -1,11 +1,9 @@
 
 import * as PushAPI from '@pushprotocol/restapi';
 import type { IMessageIPFS } from '@pushprotocol/restapi';
-import { Env } from '@pushprotocol/restapi';
 import { useCallback, useContext, useState } from 'react';
-import { Constants } from '../../config';
-import { ChatMainStateContext, ChatAndNotificationPropsContext } from '../../context';
-import type { ChatMainStateContextType } from '../../context/chatAndNotification/chat/chatMainStateContext';
+import { ChatDataContext } from '../../context';
+
 
 
 
@@ -19,10 +17,9 @@ const useFetchHistoryMessages
  = () => {
   const [error, setError] = useState<string>();
   const [loading, setLoading] = useState<boolean>(false);
-  const { chats,setChat,selectedChatId} =
-  useContext<ChatMainStateContextType>(ChatMainStateContext);
-  const { account, env,decryptedPgpPvtKey } =
-  useContext<any>(ChatAndNotificationPropsContext);
+
+  const { account, env,pgpPrivateKey } =
+  useContext<any>(ChatDataContext);
 
   const historyMessages = useCallback(async ({threadHash,limit = 10,}:HistoryMessagesParams) => {
 
@@ -31,40 +28,24 @@ const useFetchHistoryMessages
         const chatHistory:IMessageIPFS[] = await PushAPI.chat.history({
             threadhash: threadHash,
             account: account,
-            toDecrypt: decryptedPgpPvtKey ? true : false,
-            pgpPrivateKey: String(decryptedPgpPvtKey),
+            toDecrypt: pgpPrivateKey ? true : false,
+            pgpPrivateKey: String(pgpPrivateKey),
             limit: limit,
             env: env
           });
           chatHistory.reverse();
-          if (chats.get(selectedChatId as string)) {
-            const uniqueMap: { [timestamp: number]: IMessageIPFS } = {};
-            const messages = Object.values(
-              [...chatHistory, ...chats.get(selectedChatId  as string)!.messages].reduce((uniqueMap, message) => {
-                if (message.timestamp && !uniqueMap[message.timestamp]) {
-                  uniqueMap[message.timestamp] = message;
-                }
-                return uniqueMap;
-              }, uniqueMap)
-            );
-            setChat(selectedChatId  as string, {
-              messages: messages,
-              lastThreadHash: chatHistory[0].link
-            });
-          } else {
-            setChat(selectedChatId  as string, { messages: chatHistory, lastThreadHash: chatHistory[0].link });
-          }
+       return chatHistory;
     } catch (error: Error | any) {
       setLoading(false);
       setError(error.message);
       console.log(error);
+      return;
     } finally {
       setLoading(false);
     }
-  }, [chats]);
+  }, [pgpPrivateKey,account,env]);
 
   return { historyMessages, error, loading };
 };
 
-export default useFetchHistoryMessages
-;
+export default useFetchHistoryMessages;